Modify table columns in spreadsheet-data.js instead of geodata, also set tableJson
