+++ addon/templates/components/geojson-parse-data.hbs
@@ -38,7 +38,7 @@
       {{/each}}
     </select>
 {{/each}}
-<button {{action 'generateFeaturs'}}>Generate Features</button>
+<button {{action 'generateFeatures'}}>Generate Features</button>
 
 </div>
 {{/if}}
